personalityTest/requirements.txt
Babel==2.12.1
blinker==1.6.2
click==8.1.3
colorama==0.4.6
distlib==0.3.6
distro==1.7.0
et-xmlfile==1.1.0
filelock==3.12.1
Flask==2.3.2
Flask-Cors==3.0.10
Flask-Login==0.6.2
Flask-MySQLdb==1.0.1
flask-paginate==2022.1.8
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mysql-connector-python==8.0.33
mysqlclient==2.1.1
numpy==1.22.3
openpyxl==3.0.9
pandas==1.4.1
platformdirs==3.5.3
protobuf==3.20.3
PyMySQL==1.0.3
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
SQLAlchemy==2.0.16
tabula==1.0.5
tabula-py==2.3.0
tkcalendar==1.6.1
typing_extensions==4.6.3
virtualenv==20.23.0
Werkzeug==2.3.6
XlsxWriter==3.0.3
